Refactoring of the CLI module. - Added missing JavaDoc when relevant. - Changed some tools used by more efficient ones - Refactored TestView.java - Excluded TestView and its ecosystem from the target.
