+++ cli/src/main/java/com/dexesttp/hkxpack/cli/utils/WrongSizeException.java
@@ -5,5 +5,7 @@ package com.dexesttp.hkxpack.cli.utils;
  */
 public class WrongSizeException extends Exception {
 	private static final long serialVersionUID = -4241588440572992978L;
-
+	public WrongSizeException(String option) {
+		super(option);
+	}
 }
